MODEXPS-1::Create mod-data-export-spring and mod-data-export-worker modules basing on the PoC Temp local files and S3 files processing refactoring.
